chore: refactor plugin api
