+++ src/main/java/io/bonitoo/qa/conf/data/ItemConfigDeserializer.java
@@ -6,12 +6,10 @@ import com.fasterxml.jackson.databind.JsonNode;
 import io.bonitoo.qa.conf.VirDevConfigException;
 import io.bonitoo.qa.conf.VirDevDeserializer;
 import io.bonitoo.qa.data.ItemType;
-import io.bonitoo.qa.plugin.ItemPluginMill;
-import io.bonitoo.qa.plugin.PluginConfigException;
+import io.bonitoo.qa.plugin.item.ItemPluginMill;
 import io.bonitoo.qa.plugin.PluginProperties;
 import io.bonitoo.qa.plugin.PluginResultType;
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
