+++ src/main/java/io/bonitoo/qa/conf/device/DeviceConfigDeserializer.java
@@ -3,17 +3,16 @@ package io.bonitoo.qa.conf.device;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.node.NullNode;
 import io.bonitoo.qa.conf.Config;
 import io.bonitoo.qa.conf.VirDevConfigException;
 import io.bonitoo.qa.conf.VirDevDeserializer;
 import io.bonitoo.qa.conf.data.SampleConfig;
 import io.bonitoo.qa.conf.data.SampleConfigRegistry;
-import io.bonitoo.qa.plugin.SamplePlugin;
-import io.bonitoo.qa.plugin.SamplePluginConfig;
-import io.bonitoo.qa.plugin.SamplePluginConfigClass;
-import io.bonitoo.qa.plugin.SamplePluginMill;
+import io.bonitoo.qa.plugin.sample.SamplePlugin;
+import io.bonitoo.qa.plugin.sample.SamplePluginConfig;
+import io.bonitoo.qa.plugin.sample.SamplePluginConfigClass;
+import io.bonitoo.qa.plugin.sample.SamplePluginMill;
 import io.bonitoo.qa.util.LogHelper;
 import java.io.IOException;
 import java.lang.annotation.Annotation;
