+++ src/main/java/io/bonitoo/qa/device/GenericDevice.java
@@ -8,9 +8,8 @@ import io.bonitoo.qa.data.GenericSample;
 import io.bonitoo.qa.data.Sample;
 import io.bonitoo.qa.mqtt.client.MqttClientBlocking;
 import io.bonitoo.qa.plugin.PluginConfigException;
-import io.bonitoo.qa.plugin.SamplePlugin;
-import io.bonitoo.qa.plugin.SamplePluginConfig;
-import io.bonitoo.qa.plugin.SamplePluginMill;
+import io.bonitoo.qa.plugin.sample.SamplePluginConfig;
+import io.bonitoo.qa.plugin.sample.SamplePluginMill;
 import io.bonitoo.qa.util.LogHelper;
 import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
