+++ src/main/java/io/bonitoo/qa/plugin/item/ItemGenPlugin.java
@@ -1,6 +1,7 @@
-package io.bonitoo.qa.plugin;
+package io.bonitoo.qa.plugin.item;
 
 import io.bonitoo.qa.conf.data.ItemConfig;
+import io.bonitoo.qa.plugin.PluginProperties;
 
 /**
  * Point from which Item Generator Plugins will be created.
