+++ src/test/java/io/bonitoo/qa/plugin/eg/EmptyItemGenPlugin.java
@@ -1,7 +1,7 @@
 package io.bonitoo.qa.plugin.eg;
 
 import io.bonitoo.qa.conf.data.ItemConfig;
-import io.bonitoo.qa.plugin.ItemGenPlugin;
+import io.bonitoo.qa.plugin.item.ItemGenPlugin;
 import io.bonitoo.qa.plugin.PluginProperties;
 import lombok.NoArgsConstructor;
