+++ src/test/java/io/bonitoo/qa/plugin/eg/InfluxLPSamplePlugin.java
@@ -6,9 +6,9 @@ import com.fasterxml.jackson.databind.ObjectWriter;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import io.bonitoo.qa.data.Sample;
 import io.bonitoo.qa.plugin.PluginProperties;
-import io.bonitoo.qa.plugin.SamplePlugin;
-import io.bonitoo.qa.plugin.SamplePluginConfig;
-import io.bonitoo.qa.plugin.SamplePluginConfigClass;
+import io.bonitoo.qa.plugin.sample.SamplePlugin;
+import io.bonitoo.qa.plugin.sample.SamplePluginConfig;
+import io.bonitoo.qa.plugin.sample.SamplePluginConfigClass;
 import lombok.Getter;
 import lombok.Setter;
 import org.slf4j.Logger;
