+++ src/test/java/io/bonitoo/qa/plugin/eg/InfluxLPSamplePluginConf.java
@@ -2,7 +2,7 @@ package io.bonitoo.qa.plugin.eg;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.bonitoo.qa.conf.data.ItemConfig;
-import io.bonitoo.qa.plugin.SamplePluginConfig;
+import io.bonitoo.qa.plugin.sample.SamplePluginConfig;
 import lombok.Getter;
 import java.util.List;
 import java.util.Map;
